Add extra validations when adding bros.
Previously, you could add a non-existant friend, or add yourself
as a friend. Add extra validations to stop this.

Make current user's username available in the ApiRequest struct,
changing the AuthUser function call to directly modify the ApiRequest.

Add a get after a new device is "Put", to ensure that the device is
applied to the datastore before returning.